[#207] Refactor the serial init in a more logical and smaller way of doing things
